Add files via upload
Karel starts in the bottom left corner of a world with three "obstacles" (vertical walls in the bottommost row).  Your job is to write code which gets Karel over the obstacles and puts beepers in the squares directly to the right of the walls, and moves Karel to the bottom right corner of the world.