Add files via upload
Karel starts off in the corner of her house as shown in the world. She needs to collect some food, represented (as all objects in Karel's world are) by a beeper, from outside the doorway and then to return to her initial position. You can assume that every part of the world is always the same. The house is exactly this size, the door is always in the position shown, and the beeper is just outside the door. Thus, all you have to do is write the sequence of commands necessary to have Karel

Move to the beeper,

Pick it up, and

Return to her starting point.

Even though the program is only a few lines, it is still worth getting at least a little practice in decomposition. In your solution, include a function for moving to the package, and returning to the starting point.